[SW-1530] Conversion to H2OFrame needs to work without running H2O client
