+++ core/src/main/scala/org/apache/spark/h2o/converters/WriteConverterCtxUtils.scala
@@ -23,11 +23,9 @@ import org.apache.spark.h2o.backends.internal.InternalWriteConverterCtx
 import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.h2o.{H2OContext, _}
 import org.apache.spark.rdd.h2o.H2OAwareRDD
-import water.fvec.{Chunk, H2OFrame}
-import water.util.Log
-import water.{DKV, ExternalFrameUtils, Key, _}
+import water.{ExternalFrameUtils, _}
 
-import scala.collection.{immutable, mutable}
+import scala.collection.immutable
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe._
 
@@ -75,7 +73,11 @@ object WriteConverterCtxUtils {
   def convert[T: ClassTag: TypeTag](hc: H2OContext, rddInput: RDD[T], keyName: String, colNames: Array[String], expectedTypes: Array[Byte],
                  maxVecSizes: Array[Int], sparse: Array[Boolean], func: ConversionFunction[T]) = {
     // Make an H2O data Frame - but with no backing data (yet)
-    initFrame(keyName, colNames)
+    if (hc.getConf.runsInExternalClusterMode) {
+      ExternalWriteConverterCtx.initFrame(keyName, colNames)
+    } else {
+      InternalWriteConverterCtx.initFrame(keyName, colNames)
+    }
 
     // prepare required metadata based on the used backend
     val uploadPlan = if (hc.getConf.runsInExternalClusterMode) {
@@ -106,99 +108,13 @@ object WriteConverterCtxUtils {
     } else {
       expectedTypes
     }
-    val fr = new H2OFrame(finalizeFrame(keyName, res, types))
 
-    if (Log.isLoggingFor("DEBUG")) {
-      if (!fr.vecs().isEmpty) {
-        Log.debug("Number of chunks on frame: " + fr.anyVec.nChunks)
-        val nodes = mutable.Map.empty[H2ONode, Int]
-        (0 until fr.anyVec().nChunks()).foreach { i =>
-          val home = fr.anyVec().chunkKey(i).home_node()
-          if (!nodes.contains(home)) {
-            nodes += (home -> 0)
-          }
-          nodes(home) = nodes(home) + 1
-        }
-        Log.debug("Frame distributed on nodes:")
-        nodes.foreach {
-          case (node, n) =>
-            Log.debug(node + ": " + n + " chunks.")
-        }
-      }
+    if (hc.getConf.runsInExternalClusterMode) {
+      ExternalWriteConverterCtx.finalizeFrame(keyName, res, types)
+    } else {
+      InternalWriteConverterCtx.finalizeFrame(keyName, res, types)
     }
 
-    // Validate num of chunks in each vector
-    if (!fr.vecs().isEmpty) {
-      val first = fr.vecs.head
-      fr.vecs.tail.foreach { vec =>
-        if (vec.nChunks() != first.nChunks()) {
-          throw new IllegalArgumentException(
-            s"""
-               | Vectors have different number of chunks: ${fr.vecs().map(_.nChunks()).mkString(", ")}
-        """.stripMargin)
-        }
-      }
-    }
-
-    // Validate that espc is the same in each vector
-    if (!fr.vecs().isEmpty) {
-      val layouts = fr.vecs().map(_.espc())
-      val first = layouts.head
-      layouts.tail.foreach { espc =>
-        if (first != espc) {
-          throw new IllegalArgumentException(
-            s"""
-               | Invalid shape of H2O Frame:
-               |
-          | ${layouts.zipWithIndex.map { case (arr, idx) => s"Vec $idx has layout: ${arr.mkString(", ")}\n" }}
-               |
-        }
-        """.stripMargin)
-        }
-      }
-    }
-
-    // Check number of entries in each chunk
-    if (!fr.vecs().isEmpty) {
-
-      new MRTask() {
-
-        override def map(cs: Array[Chunk]): Unit = {
-
-          val values = cs.map(_.len())
-          if (values.length > 1) {
-            val firstLen = values.head
-            values.tail.zipWithIndex.foreach { case (len, idx) =>
-              if (firstLen != len) {
-                throw new IllegalArgumentException(s"Chunks have different sizes in different vectors: $firstLen in vector 1 and" +
-                  s"$len in vector $idx")
-              }
-            }
-          }
-        }
-
-      }.doAll(fr)
-
-    }
-
-    fr
-  }
-
-  private def initFrame(keyName: String, names: Array[String]): Unit = {
-    val fr = new water.fvec.Frame(Key.make[Frame](keyName))
-    water.fvec.FrameUtils.preparePartialFrame(fr, names)
-    // Save it directly to DKV
-    fr.update()
-  }
-
-  private def finalizeFrame(keyName: String,
-                            res: Array[Long],
-                            colTypes: Array[Byte],
-                            colDomains: Array[Array[String]] = null): Frame = {
-
-    val fr = DKV.getGet[Frame](keyName)
-    water.fvec.FrameUtils.finalizePartialFrame(fr, res, colDomains, colTypes)
-    fr
   }
 
 }
